moved configurer to service package top level
