+++ main.go
@@ -19,7 +19,7 @@ import (
 	"github.com/giantswarm/draughtsman/service"
 	"github.com/giantswarm/draughtsman/service/deployer"
 	"github.com/giantswarm/draughtsman/service/deployer/eventer/github"
-	"github.com/giantswarm/draughtsman/service/deployer/installer/configurer/configmap"
+	"github.com/giantswarm/draughtsman/service/configurer/configmap"
 	"github.com/giantswarm/draughtsman/service/deployer/installer/helm"
 	slacknotifier "github.com/giantswarm/draughtsman/service/deployer/notifier/slack"
 	slackspec "github.com/giantswarm/draughtsman/slack"
